// web/src/types/test.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type AmateurRadioTest = {
  [k: string]: unknown;
} & {
  /**
   * Unique test ID in format test_YYYYMMDD_HHMMSS
   */
  test_id: string;
  /**
   * ISO 8601 timestamp of test creation
   */
  timestamp: string;
  source_pool: {
    license_class: "technician" | "general" | "extra";
    /**
     * Question pool version years
     */
    version: string;
    /**
     * Source question pool filename
     */
    file: string;
    [k: string]: unknown;
  };
  questions: {
    /**
     * Question ID matching license class
     */
    id: string;
    /**
     * Question group identifier
     */
    group: string;
    /**
     * The question text
     */
    question: string;
    /**
     * @minItems 4
     * @maxItems 4
     */
    answers: [
      {
        option: "A" | "B" | "C" | "D";
        text: string;
        [k: string]: unknown;
      },
      {
        option: "A" | "B" | "C" | "D";
        text: string;
        [k: string]: unknown;
      },
      {
        option: "A" | "B" | "C" | "D";
        text: string;
        [k: string]: unknown;
      },
      {
        option: "A" | "B" | "C" | "D";
        text: string;
        [k: string]: unknown;
      }
    ];
    /**
     * Must match one of the answer options
     */
    correct_answer: "A" | "B" | "C" | "D";
    [k: string]: unknown;
  }[];
  [k: string]: unknown;
};